+++ tests/testthat/test-write.R
@@ -136,3 +136,37 @@ test_that("writeH5AD DelayedArray layer works", {
 
     expect_identical(counts(sce), assay(out, "layer"))
 })
+
+test_that("writeH5AD works with colData list columns", {
+
+    list_sce <- sce
+    colData(list_sce)$ListCol <- lapply(seq_len(ncol(list_sce)), function(x) {
+        sample(LETTERS, 2)
+    })
+
+    temp <- tempfile(fileext = ".h5ad")
+
+    expect_warning(writeH5AD(list_sce, temp), "colData columns are not atomic")
+    expect_true(file.exists(temp))
+
+    # Knowing what comes back is hard so just check there is something
+    out <- readH5AD(temp)
+    expect_true("ListCol" %in% names(metadata(out)$.colData))
+})
+
+test_that("writeH5AD works with rowData list columns", {
+
+    list_sce <- sce
+    rowData(list_sce)$ListCol <- lapply(seq_len(nrow(list_sce)), function(x) {
+        sample(LETTERS, 2)
+    })
+
+    temp <- tempfile(fileext = ".h5ad")
+
+    expect_warning(writeH5AD(list_sce, temp), "rowData columns are not atomic")
+    expect_true(file.exists(temp))
+
+    # Knowing what comes back is hard so just check there is something
+    out <- readH5AD(temp)
+    expect_true("ListCol" %in% names(metadata(out)$.rowData))
+})